+++ docs/history.rst
@@ -3,6 +3,12 @@
 Release History
 ===============
 
+`Next Release`_
+---------------
+- Make the platform directory name an option for ``prep-it``
+- Add ``--sleep`` parameter to ``wait-for``
+- Normalize parameter processing between commands
+
 `1.0.0`_ (2016-01-27)
 ---------------------
 - Infect the world!