+++ docs/history.rst
@@ -8,6 +8,8 @@ Release History
 - Make the platform directory name an option for ``prep-it``
 - Add ``--sleep`` parameter to ``wait-for``
 - Normalize parameter processing between commands
+- Do not create a postgres database if the database-specific
+  environment variable exists.
 
 `1.0.0`_ (2016-01-27)
 ---------------------